Move step tests to designated folder to match folder layout in src
